+++ framework/domain/src/test/java/is/codion/framework/domain/entity/TypeReferenceTest.java
@@ -1,0 +1,36 @@
+/*
+ * This file is part of Codion.
+ *
+ * Codion is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Codion is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Codion.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Copyright (c) 2024, Björn Darri Sigurðsson.
+ */
+package is.codion.framework.domain.entity;
+
+import is.codion.framework.domain.entity.attribute.TypeReference;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class TypeReferenceTest {
+
+	@Test
+	void test() {
+		new TypeReference<List<Integer>>() {};
+		assertThrows(IllegalArgumentException.class, () -> new TypeReference<Integer>() {});
+	}
+}