+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.17.42
 ### is.codion.framework.domain
 - DefaultEntity refactored.
+- TypeReference now throws exception if used to specify a non-parameterized type.
 
 ## 0.17.41
 ### is.codion.common.core